# api/swagger.json
{
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Documentation for My API.",
    "title": "My API.",
    "version": "1.0.0"
  },
  "host": "localhost:8081",
  "paths": {
...
    "/api/v1/files": {
      "get": {
        "tags": [
          "files"
        ],
        "summary": "Список всех файлов.",
        "operationId": "listFiles",
        "responses": {
          "200": {
            "$ref": "#/responses/getFileResponse"
          },
          "500": {
            "$ref": "#/responses/errorResponse"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "files"
        ],
        "summary": "Загрузить новый файл.",
        "operationId": "createFile",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateFileRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/responses/getFileResponse"
          },
          "400": {
            "$ref": "#/responses/errorResponse"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/files/{id}": {
      "get": {
        "tags": [
          "files"
        ],
        "summary": "Получить файл по ID.",
        "operationId": "getFile",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/getFileResponse"
          },
          "404": {
            "$ref": "#/responses/errorResponse"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "files"
        ],
        "summary": "Обновить метаданные файла по ID.",
        "operationId": "updateFile",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/getFileResponse"
          },
          "400": {
            "$ref": "#/responses/errorResponse"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "files"
        ],
        "summary": "Удалить файл по ID.",
        "operationId": "deleteFile",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "500": {
            "$ref": "#/responses/errorResponse"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/api/v1/products": {
      "get": {
        "tags": [
          "products"
        ],
        "summary": "Список всех продуктов.",
        "operationId": "listProducts",
        "responses": {
          "200": {
            "$ref": "#/responses/getProductResponse"
          },
          "500": {
            "$ref": "#/responses/errorResponse"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "products"
        ],
        "summary": "Создать новый продукт.",
        "operationId": "createProduct",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateProductRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/responses/getProductResponse"
          },
          "400": {
            "$ref": "#/responses/errorResponse"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/products/{id}": {
      "get": {
        "tags": [
          "products"
        ],
        "summary": "Получить продукт по ID.",
        "operationId": "getProduct",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/getProductResponse"
          },
          "404": {
            "$ref": "#/responses/errorResponse"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "products"
        ],
        "summary": "Обновить продукт по ID.",
        "operationId": "updateProduct",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/getProductResponse"
          },
          "400": {
            "$ref": "#/responses/errorResponse"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "products"
        ],
        "summary": "Удалить продукт по ID.",
        "operationId": "deleteProduct",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "500": {
            "$ref": "#/responses/errorResponse"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/settings": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Список всех настроек.",
        "operationId": "listSettings",
        "responses": {
          "200": {
            "$ref": "#/responses/getSettingResponse"
          },
          "500": {
            "$ref": "#/responses/errorResponse"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Создать новую настройку.",
        "operationId": "createSetting",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateSettingRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/responses/getSettingResponse"
          },
          "400": {
            "$ref": "#/responses/errorResponse"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v1/settings/{id}": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Получить настройку по ID.",
        "operationId": "getSetting",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/getSettingResponse"
          },
          "404": {
            "$ref": "#/responses/errorResponse"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Обновить настройку по ID.",
        "operationId": "updateSetting",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/getSettingResponse"
          },
          "400": {
            "$ref": "#/responses/errorResponse"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "settings"
        ],
        "summary": "Удалить настройку по ID.",
        "operationId": "deleteSetting",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "500": {
            "$ref": "#/responses/errorResponse"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "CreateFileRequest": {
      "description": "CreateFileRequest represents the request body for uploading a file",
      "type": "object",
      "properties": {
        "client_id": {
          "description": "Client ID",
          "type": "string",
          "x-go-name": "ClientID",
          "example": "550e8400-e29b-41d4-a716-446655440000"
        },
        "name": {
          "description": "File name",
          "type": "string",
          "x-go-name": "Name",
          "example": "contract.pdf"
        },
        "storage_reference": {
          "description": "Storage reference path",
          "type": "string",
          "x-go-name": "StorageReference",
          "example": "uploads/2023/10/file.pdf"
        }
      },
      "x-go-package": "github.com/Ozenkol/rbk-go-final/internal/delivery/http/requests"
    },
...
    "CreateProductRequest": {
      "description": "CreateProductRequest represents the request body for creating a product",
      "type": "object",
      "properties": {
        "description": {
          "description": "Product description",
          "type": "string",
          "x-go-name": "Description",
          "example": "Access to all premium features"
        },
        "name": {
          "description": "Product name",
          "type": "string",
          "x-go-name": "Name",
          "example": "Premium Subscription"
        },
        "price": {
          "$ref": "#/definitions/Money"
        },
        "thumbnail": {
          "$ref": "#/definitions/StorageReference"
        }
      },
      "x-go-package": "github.com/Ozenkol/rbk-go-final/internal/delivery/http/requests"
    },
    "CreateSettingRequest": {
      "description": "CreateSettingRequest represents the request body for creating a setting",
      "type": "object",
      "properties": {
        "key": {
          "description": "Setting key",
          "type": "string",
          "x-go-name": "Key",
          "example": "maintenance_mode"
        },
        "value": {
          "description": "Setting value",
          "type": "string",
          "x-go-name": "Value",
          "example": "false"
        }
      },
      "x-go-package": "github.com/Ozenkol/rbk-go-final/internal/delivery/http/requests"
    },
...
    "Currency": {
      "type": "object",
      "properties": {
        "Code": {
          "type": "string"
        },
        "Name": {
          "type": "string"
        }
      },
      "x-go-package": "github.com/Ozenkol/rbk-go-final/internal/domain/shared"
    },
...
    "File": {
      "type": "object",
      "properties": {
        "ClientID": {
          "type": "string"
        },
        "CompanyID": {
          "type": "string"
        },
        "ID": {
          "type": "string"
        },
        "Name": {
          "type": "string"
        },
        "StorageReference": {
          "type": "string"
        },
        "UserID": {
          "type": "string"
        }
      },
      "x-go-package": "github.com/Ozenkol/rbk-go-final/internal/domain/file"
    },
...
    "Money": {
      "type": "object",
      "properties": {
        "Amount": {
          "type": "number",
          "format": "double"
        },
        "Currency": {
          "$ref": "#/definitions/Currency"
        }
      },
      "x-go-package": "github.com/Ozenkol/rbk-go-final/internal/domain/shared"
    },
...
    "Product": {
      "type": "object",
      "properties": {
        "CompanyID": {
          "type": "string"
        },
        "Description": {
          "type": "string"
        },
        "ID": {
          "type": "string"
        },
        "Name": {
          "type": "string"
        },
        "Price": {
          "$ref": "#/definitions/Money"
        },
        "Thumbnail": {
          "$ref": "#/definitions/StorageReference"
        },
        "UserID": {
          "type": "string"
        }
      },
      "x-go-package": "github.com/Ozenkol/rbk-go-final/internal/domain/product"
    },
    "Setting": {
      "type": "object",
      "properties": {
        "CompanyID": {
          "type": "string"
        },
        "ID": {
          "type": "string"
        },
        "Key": {
          "type": "string"
        },
        "UserID": {
          "type": "string"
        },
        "Value": {
          "type": "string"
        }
      },
      "x-go-package": "github.com/Ozenkol/rbk-go-final/internal/domain/setting"
    },
    "StorageReference": {
      "type": "object",
      "properties": {
        "ServiceName": {
          "type": "string"
        },
        "URL": {
          "type": "string"
        }
      },
      "x-go-package": "github.com/Ozenkol/rbk-go-final/internal/domain/shared"
    },
...
  },
  "responses": {
...
    "errorResponse": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "x-go-name": "Error"
          }
        }
      }
    },
...
    "getFileResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/File"
      }
    },
...
    "getProductResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/Product"
      }
    },
    "getSettingResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/Setting"
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}