# internal/delivery/http/handlers/setting_handler.go
package handlers

import (
	"log/slog"
	"net/http"

	"github.com/Ozenkol/rbk-go-final/internal/application/command"
	"github.com/Ozenkol/rbk-go-final/internal/application/query"
	http_deps "github.com/Ozenkol/rbk-go-final/internal/delivery/http/deps"
	http_requests "github.com/Ozenkol/rbk-go-final/internal/delivery/http/requests"
	"github.com/Ozenkol/rbk-go-final/internal/domain/setting"
	"github.com/gin-gonic/gin"
)

type SettingHandler struct {
	deps *http_deps.Dependencies
	logs *slog.Logger
}

func NewSettingHandler(deps *http_deps.Dependencies, logs *slog.Logger) *SettingHandler {
	return &SettingHandler{deps: deps, logs: logs}
}

// swagger:route POST /api/v1/settings settings createSetting
// Создать новую настройку.
// Security:
//   Bearer:
// responses:
//   201: getSettingResponse
//   400: errorResponse
func (h *SettingHandler) Create(c *gin.Context) {
	var req http_requests.CreateSettingRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	token := c.GetHeader("Authorization")
	userID, companyID, err := h.deps.App.Services.AuthService.GetAuthInfoFromToken(token)
	if err != nil {
		c.JSON(http.StatusUnauthorized, gin.H{"error": "Unauthorized"})
		return
	}

	s := &setting.Setting{
		UserID:    userID,
		CompanyID: companyID,
		Key:       req.Key,
		Value:     req.Value,
	}

	res, err := h.deps.App.Commands.CreateSetting.Handle(c.Request.Context(), command.CreateSettingCommand{
		Setting: s,
	})
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusCreated, res)
}

// swagger:route GET /api/v1/settings/{id} settings getSetting
// Получить настройку по ID.
// Security:
//   Bearer:
// responses:
//   200: getSettingResponse
//   404: errorResponse
func (h *SettingHandler) GetByID(c *gin.Context) {
	id := c.Param("id")
	res, err := h.deps.App.Queries.GetSettingByID.Handle(c.Request.Context(), query.FetchSettingByID{ID: id})
	if err != nil {
		c.JSON(http.StatusNotFound, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusOK, res)
}

// swagger:route GET /api/v1/settings settings listSettings
// Список всех настроек.
// Security:
//   Bearer:
// responses:
//   200: []getSettingResponse
//   500: errorResponse
func (h *SettingHandler) List(c *gin.Context) {
	res, err := h.deps.App.Queries.ListSettings.Handle(c.Request.Context(), query.FetchSettingList{})
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusOK, res)
}

// swagger:route PUT /api/v1/settings/{id} settings updateSetting
// Обновить настройку по ID.
// Security:
//   Bearer:
// responses:
//   200: getSettingResponse
//   400: errorResponse
func (h *SettingHandler) Update(c *gin.Context) {
	id := c.Param("id")
	var req http_requests.UpdateSettingRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	token := c.GetHeader("Authorization")
	userID, companyID, err := h.deps.App.Services.AuthService.GetAuthInfoFromToken(token)
	if err != nil {
		c.JSON(http.StatusUnauthorized, gin.H{"error": "Unauthorized"})
		return
	}

	s := &setting.Setting{
		ID:        id,
		UserID:    userID,
		CompanyID: companyID,
		Key:       req.Key,
		Value:     req.Value,
	}

	res, err := h.deps.App.Commands.UpdateSetting.Handle(c.Request.Context(), command.UpdateSettingCommand{
		Setting: s,
	})
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusOK, res)
}

// swagger:route DELETE /api/v1/settings/{id} settings deleteSetting
// Удалить настройку по ID.
// Security:
//   Bearer:
// responses:
//   204:
//   500: errorResponse
func (h *SettingHandler) Delete(c *gin.Context) {
	id := c.Param("id")
	err := h.deps.App.Commands.DeleteSetting.Handle(c.Request.Context(), command.DeleteSettingCommand{ID: id})
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
		return
	}
	c.JSON(http.StatusNoContent, nil)
}